+++ device/api/umd/device/tt_device/blackhole_tt_device.h
@@ -10,7 +10,7 @@
 #include <stdexcept>
 
 #include "umd/device/tlb.h"
-#include "umd/device/tt_device.h"
+#include "umd/device/tt_device/tt_device.h"
 
 namespace tt::umd {
