+++ device/cluster.cpp
@@ -42,7 +42,7 @@
 #include "umd/device/tlb.h"
 #include "umd/device/tt_arch_types.h"
 #include "umd/device/tt_cluster_descriptor.h"
-#include "umd/device/tt_device.h"
+#include "umd/device/tt_device/tt_device.h"
 #include "yaml-cpp/yaml.h"
 
 using namespace boost::interprocess;
