+++ tests/blackhole/test_silicon_driver_bh.cpp
@@ -14,7 +14,7 @@
 #include "tests/test_utils/device_test_utils.hpp"
 #include "tests/test_utils/generate_cluster_desc.hpp"
 #include "umd/device/tt_cluster_descriptor.h"
-#include "umd/device/tt_device_blackhole.h"
+#include "umd/device/tt_device/blackhole_tt_device.h"
 
 using namespace tt::umd;
